requirements.txt
anyio==4.8.0
certifi==2024.12.14
document==1.0
h11==0.14.0
httpcore==1.0.7
httpx==0.28.1
idna==3.10
lxml==5.3.0
python-docx==1.1.0
python-telegram-bot==20.7
sniffio==1.3.1
typing_extensions==4.12.2
Pillow==10.2.0
PyMuPDF==1.24.5
